# form/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetReport": {
        "description": "agentd -> Form -> analyzer: one host, one collection cycle.",
        "properties": {
          "assets": {
            "items": {
              "discriminator": {
                "mapping": {
                  "account": "#/components/schemas/Account",
                  "container": "#/components/schemas/Container",
                  "credential": "#/components/schemas/Credential",
                  "image": "#/components/schemas/Image",
                  "package": "#/components/schemas/Package",
                  "port": "#/components/schemas/Port",
                  "security_product": "#/components/schemas/SecurityProduct",
                  "service": "#/components/schemas/Service"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Package"
                },
                {
                  "$ref": "#/components/schemas/Service"
                },
                {
                  "$ref": "#/components/schemas/Port"
                },
                {
                  "$ref": "#/components/schemas/Account"
                },
                {
                  "$ref": "#/components/schemas/Credential"
                },
                {
                  "$ref": "#/components/schemas/Container"
                },
                {
                  "$ref": "#/components/schemas/Image"
                },
                {
                  "$ref": "#/components/schemas/SecurityProduct"
                }
              ]
            },
            "maxItems": 4096,
            "title": "Assets",
            "type": "array"
          },
          "collected_at": {
            "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "detector_runs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetectorRun"
                },
                "maxItems": 32,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detectors explicitly executed by the producer. null means legacy/unknown; an empty list means the producer confirms none were enabled.",
            "title": "Detector Runs"
          },
          "host": {
            "$ref": "#/components/schemas/HostInfo"
          },
          "report_id": {
            "maxLength": 256,
            "title": "Report Id",
            "type": "string"
          },
          "scanner_version": {
            "maxLength": 256,
            "title": "Scanner Version",
            "type": "string"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticated Agent identity injected by Form; never trusted from payload",
            "title": "Source Agent Id"
          },
          "source_target_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Form-owned registered target attribution; absent only for unbound legacy telemetry",
            "title": "Source Target Id"
          },
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/Vulnerability"
            },
            "maxItems": 4096,
            "title": "Vulnerabilities",
            "type": "array"
          }
        },
        "required": [
          "report_id",
          "collected_at",
          "scanner_version",
          "host"
        ],
        "title": "AssetReport",
        "type": "object"
      },
...
      "CapabilityGraph": {
        "description": "external exporter -> Form -> analyzer: technique capability graph + templates.",
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/TechniqueCapability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "exported_at": {
            "anyOf": [
              {
                "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exported At"
          },
          "ontology_version": {
            "description": "Shared fact-vocabulary version",
            "title": "Ontology Version",
            "type": "string"
          },
          "source": {
            "default": "",
            "description": "Producing tool (opaque label)",
            "title": "Source",
            "type": "string"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/AttackTemplate"
            },
            "title": "Templates",
            "type": "array"
          }
        },
        "required": [
          "ontology_version"
        ],
        "title": "CapabilityGraph",
        "type": "object"
      },
...
      "GuardEventBatch": {
        "description": "agent-respond -> Form -> analyzer: protection events from one host.",
        "properties": {
          "agent_version": {
            "maxLength": 256,
            "title": "Agent Version",
            "type": "string"
          },
          "batch_id": {
            "maxLength": 256,
            "title": "Batch Id",
            "type": "string"
          },
          "collected_at": {
            "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "events": {
            "items": {
              "discriminator": {
                "mapping": {
                  "fim": "#/components/schemas/FileIntegrityEvent",
                  "ids": "#/components/schemas/IdsEvent",
                  "malware": "#/components/schemas/MalwareEvent",
                  "network": "#/components/schemas/NetworkEvent",
                  "process": "#/components/schemas/ProcessEvent"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/FileIntegrityEvent"
                },
                {
                  "$ref": "#/components/schemas/MalwareEvent"
                },
                {
                  "$ref": "#/components/schemas/ProcessEvent"
                },
                {
                  "$ref": "#/components/schemas/NetworkEvent"
                },
                {
                  "$ref": "#/components/schemas/IdsEvent"
                }
              ]
            },
            "maxItems": 4096,
            "title": "Events",
            "type": "array"
          },
          "host_id": {
            "maxLength": 256,
            "title": "Host Id",
            "type": "string"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticated Agent identity injected by Form; never trusted from payload",
            "title": "Source Agent Id"
          },
          "source_target_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Form target bound to source_agent_id; absent for legacy data",
            "title": "Source Target Id"
          }
        },
        "required": [
          "batch_id",
          "collected_at",
          "host_id",
          "agent_version"
        ],
        "title": "GuardEventBatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScanScheduleInput": {
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/ScanCapability"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "interval_minutes": {
            "maximum": 43200.0,
            "minimum": 1.0,
            "title": "Interval Minutes",
            "type": "integer"
          },
          "options": {
            "$ref": "#/components/schemas/ScanJobOptions"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "target_id",
          "capability",
          "interval_minutes"
        ],
        "title": "ScanScheduleInput",
        "type": "object"
      },
...
      "TraceBatch": {
        "description": "agentd -> Form -> analyzer: trace events from one collector instance.\n\nCarries three homogeneous streams from one eBPF collection cycle: network\ntraces (5-tuple flows), file operations, and process lifecycle events.",
        "properties": {
          "batch_id": {
            "maxLength": 256,
            "title": "Batch Id",
            "type": "string"
          },
          "collected_at": {
            "description": "UTC timestamp encoded as RFC 3339 / ISO 8601",
            "format": "date-time",
            "title": "Collected At",
            "type": "string"
          },
          "collector_id": {
            "maxLength": 256,
            "title": "Collector Id",
            "type": "string"
          },
          "collector_version": {
            "maxLength": 256,
            "title": "Collector Version",
            "type": "string"
          },
          "events": {
            "description": "Network traces (5-tuple flows + IOC matches).",
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "maxItems": 4096,
            "title": "Events",
            "type": "array"
          },
          "file_events": {
            "description": "File-system operations observed by the eBPF tracer.",
            "items": {
              "$ref": "#/components/schemas/FileTraceEvent"
            },
            "maxItems": 4096,
            "title": "File Events",
            "type": "array"
          },
          "process_events": {
            "description": "Process exec/exit events observed by the eBPF tracer.",
            "items": {
              "$ref": "#/components/schemas/ProcessTraceEvent"
            },
            "maxItems": 4096,
            "title": "Process Events",
            "type": "array"
          },
          "source_agent_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticated Agent identity injected by Form; never trusted from payload",
            "title": "Source Agent Id"
          },
          "source_target_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Form-owned registered target attribution; absent only for unbound legacy telemetry",
            "title": "Source Target Id"
          }
        },
        "required": [
          "batch_id",
          "collected_at",
          "collector_id",
          "collector_version"
        ],
        "title": "TraceBatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Control plane and the only integration boundary for admin, analyzer, and agent.",
    "title": "kcatta Form",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/attack-paths": {
      "get": {
        "operationId": "proxy_attack_paths_root_attack_paths_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Attack Paths Root"
      }
    },
    "/attack-paths/{path}": {
      "get": {
        "operationId": "proxy_attack_paths_attack_paths__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Attack Paths"
      }
    },
...
    "/detect/{path}": {
      "post": {
        "operationId": "proxy_detect_detect__path__post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Detect"
      }
    },
...
    "/ingest/asset-report": {
      "post": {
        "operationId": "ingest_asset_report_ingest_asset_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Asset Report"
      }
    },
    "/ingest/capability-graph": {
      "post": {
        "operationId": "ingest_capability_graph_ingest_capability_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityGraph"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Capability Graph"
      }
    },
    "/ingest/guard-event": {
      "post": {
        "operationId": "ingest_guard_event_ingest_guard_event_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardEventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Guard Event"
      }
    },
    "/ingest/trace-batch": {
      "post": {
        "operationId": "ingest_trace_batch_ingest_trace_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Trace Batch"
      }
    },
...
    "/reports/{path}": {
      "get": {
        "operationId": "get_reports_reports__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Reports"
      },
      "post": {
        "operationId": "post_reports_reports__path__post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Reports"
      }
    },
...
    "/schedules": {
      "get": {
        "operationId": "list_schedules_schedules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScanSchedule"
                  },
                  "title": "Response List Schedules Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Schedules",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanScheduleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSchedule"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/schedules/{schedule_id}": {
      "delete": {
        "operationId": "delete_schedule_schedules__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Schedule",
        "tags": [
          "schedules"
        ]
      },
      "get": {
        "operationId": "get_schedule_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSchedule"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
...
  }
}